Reorganize file for activity form panels
